airflow/requirements.txt
praw==7.7.1
pymongo==4.5.0
minio==7.1.16
sqlparse==0.2.4
pandas
confluent_kafka==2.2.0
cassandra-driver==3.28.0